server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
chromadb

# Embeddings
sentence-transformers

# PDF Parsing
PyMuPDF  


# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests

# Logging (optional but recommended)
loguru

# File Uploads
aiofiles

# OCR
unstructured[local-inference] # Esto instala unstructured con todas las dependencias para PDFs y OCR
pdf2image
Pillow

# pip install -r requirements.txt